LTW-flask/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Bcrypt==1.0.1
cloudinary==1.32.0
python-dotenv==1.0.0
email-validator==1.3.1
pillow>=10.3.0
Flask-Migrate==4.0.4
pyodbc==4.0.35
python-slugify==8.0.1
Flask-Session==0.4.0